Create the layout of the whole Application
Form and Preview components are added for the App component. Furthermore, a context provider was wrapped around the App component in order for the child components to have access
on the formData state and a couple of event handlers. Additionally, a handlePrint function was also created within the App component and have it passed to the Form component as a prop
for it to have access to the print feature. Lastly, stylings were also applied for the overall layout of the Application.